Not really digging this mixin
